test: move assign-to-import tests to their own file
runtime-transpiler.test.ts has two pre-existing release-only failures
('// @bun > async transpiler' / '// @bun > require()') where .hbs is
parsed as JSX under the locally built release test runner but not under
debug or 'bun -e'. Those are unrelated to this change and fail with or
without it, so keep this PR's tests in a dedicated file that is clean
in every profile.
